Move child task cancellation to TaskManager
Today TransportCancelTasksAction takes care of eventually cancelling
children tasks when needed, by banning any new task from being started
using the id of the cancelled task as parent task id. Also, any ongoing
task with the parent task id set to the id of the cancelled task is
cancelled.

This change moves this logic, without modifying any behaviour, to from
TransportCancelTasksAction to TaskManager. This will allow to support
recursive cancellation of tasks as a folow-up, as currently only
children tasks are cancelled but for instance grand-children don't. Once
TaskManager is able to send ban request autonomously, it can support
cancelling grand-children tasks too.
