+++ server/src/test/java/org/elasticsearch/action/admin/cluster/node/tasks/TaskManagerTestCase.java
@@ -177,7 +177,7 @@ public abstract class TaskManagerTestCase extends ESTestCase {
                 @Override
                 protected TaskManager createTaskManager(Settings settings, ThreadPool threadPool, Set<String> taskHeaders) {
                     if (MockTaskManager.USE_MOCK_TASK_MANAGER_SETTING.get(settings)) {
-                        return new MockTaskManager(settings, threadPool, taskHeaders);
+                        return new MockTaskManager(settings, threadPool, taskHeaders, this);
                     } else {
                         return super.createTaskManager(settings, threadPool, taskHeaders);
                     }
